+++ run_prerelease.sh
@@ -1,3 +1,7 @@
 
 cp models/_original_models.json models/models.json
 cp models/_original_formats.json models/formats.json
+
+rm build/*
+rm dist/*
+rm -rf sibila.egg-info/